Pnm max val log2 (#1312)
* pnm: Reject non-standard `max_val` values, and support max_val = 1

In some case we want to reject values that are not precisely: `(1 <<
nbits) - 1`. Also reject pbm for now.

* Fix a minor documentation issue

* Remove `bits_per_sample == 1` check for now

Since we don't support packed 1-bit anyway, and atm supporting it
wouldn't help much to save memory because we'll internally store it as
3x32-bit floats anyway, I'm inclined to say that we should just assume
things are not packed in ConvertFromExternal, i.e. treat 1-bit just like
how we treat 2-bit.

Co-authored-by: Jon Sneyers <jon@cloudinary.com>