Refactor the low memory rendering pipeline and fix patch + noise combination. (#1402)
* Refactor the low memory rendering pipeline.

The code should now hopefully be more clear and undestandable.

As a nice side effect, this remove a potential race condition and
optimizes rendering of frames partially outside the current image.

* Fix the patch stage when combined with noise.

The patch stage declared to affect the additional noise channels (which
it doesn't). Because of this, the pipeline execution got confused about
the amount of y axis padding that the stage required (as it assumes that
it is the same for all channels for which it matters), resulting in
bogus values for the y position and, thus, in patches being moved by 2
pixels up.