encoder api: error codes (#823)
Adds JxlEncoderGetError which can be used to check the reason why JXL_ENCODE_ERROR was returned: it could be either bad API usage, so an application bug, or an error condition that is not caused by an application bug and an application might want to react to it, e.g. by retrying JPEG recompression without storing the jbrd or by informing the user that the input is invalid.